Remove TestContextLifecycle in favor of ApplicationContextLifecycle
Now we spawn multiple ApplicationContextLifecycle instances but the one
spawned from the ServiceLoader is the one that will be used as the root
lifecycle while the others are lifecycles around test contexts
